reorder steps
The plot_im_low and plot_im_high steps need the fits images to make the diagnostic plots, but these have been moved away in the previous steps.  I've moved the plot_im steps to before the mv steps. This also means the diagnostic images are created immediately after the image (useful if something crashes in the subtract steps)