Fix bug in getfilesize() in concat_MS_CWL.py
The function getfilesize() in concat_MS_CWL.py uses `du` to determine
the size of a given "file", in casu the size of an MS-directory.
The problem is that `du` doesn't follow symbolic links. In some cases,
however, `cwltool` creates a symbolic link to an MS-directory, instead
of copying it. This fix lets `du` follow symbolic links, so that the
results of getfilesize() are also reliable in those cases.